simplify
Was slipping into perfectionism again with trying to
structure things as well as possible. Will do it when the time
comes.
For now, a simple switch construct works.

Added some tests in bash. No need to use python unittest or rspec
or a fancy test framework. All I need is simple asserts.
So, roll my own, until I need way more.

TODO: There is a bug in the way conditionals work. Fix it